+++ src/Events/EscolaLmsCartOrderSuccessTemplateEvent.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace EscolaLms\Cart\Events;
+
+class EscolaLmsCartOrderSuccessTemplateEvent extends EscolaLmsCartTemplateEvent
+{
+}
